Safari-611.1.10/LayoutTests/svg/animations/reinserting-svg-into-document-expected.txt

Reinserting SVG animation into document should continue the animation

PASS rect.x.animVal.value is 30
PASS rect.x.animVal.value is 60
